script that deletes all State objects with a name containing the letter a

+++ 0x0F-python-object_relational_mapping/13-model_state_delete_a.py
@@ -15,7 +15,8 @@ if __name__ == "__main__":
     states_with_a = session.query(State).filter(
             State.name.like('%a%')).order_by(State.id).all()
 
-    for state in states_with_a:
-        session.delete(state)
+    if states_with_a:
+        for state in states_with_a:
+            session.delete(state)
         session.commit()
     session.close()
